requirements.txt
requests==2.26.0
python_telegram_bot==13.7
python-dotenv==0.19.1